Move state machine traits to their own module
